requirements.txt
streamlit>=1.32.2
google-generativeai>=0.7.2
openai>=1.35.0
anthropic>=0.31.2
groq>=0.9.0
python-dotenv>=1.0.1
cohere>=5.5.0
mistralai>=1.0.0
huggingface_hub>=0.23.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-cohere>=0.1.0
langchain-mistralai>=0.1.0